rename ggml-cpu-hbm
